openenv_customer_support/requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
openenv
python-dotenv